+++ ToDew/ModEntry.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright 2020 Jamie Taylor
 using System;
 using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewModdingAPI.Utilities;
@@ -37,6 +38,22 @@ namespace ToDew {
                 api.RegisterModConfig(ModManifest, () => config = new ModConfig(), () => Helper.WriteConfig(config));
                 api.RegisterSimpleOption(ModManifest, "Hotkey", "The key to bring up the to-do list", () => config.hotkey, (SButton val) => config.hotkey = val);
                 api.RegisterSimpleOption(ModManifest, "Debug", "Enable debugging output in the log", () => config.debug, (bool val) => config.debug = val);
+            }
+
+            // integrate with MobilePhone, if installed
+            var phoneApi = Helper.ModRegistry.GetApi<IMobilePhoneApi>("aedenthorn.MobilePhone");
+            if (phoneApi != null) {
+                // This is a whole lot of trouble to be able to use something out of one of the built-in
+                // tile sheets, since the mobile phone api doesn't support specifying a rectangle for
+                // the sprite.
+                Texture2D originalTexture = Game1.content.Load<Texture2D>("Maps\\TownIndoors");
+                Rectangle sourceRectangle = new Rectangle(202, 1870, 48, 48);
+                Texture2D cropTexture = new Texture2D(Game1.graphics.GraphicsDevice, sourceRectangle.Width, sourceRectangle.Height);
+                Color[] data = new Color[sourceRectangle.Width * sourceRectangle.Height];
+                originalTexture.GetData(0, sourceRectangle, data, 0, data.Length);
+                cropTexture.SetData(data);
+
+                phoneApi.AddApp(Helper.ModRegistry.ModID, "To-Dew", () => { Game1.activeClickableMenu = new ToDoMenu(this.Monitor, this.list); } , cropTexture);
             }
         }
 
@@ -79,5 +96,9 @@ namespace ToDew {
         void RegisterModConfig(IManifest mod, Action revertToDefault, Action saveToFile);
         void RegisterSimpleOption(IManifest mod, string optionName, string optionDesc, Func<bool> optionGet, Action<bool> optionSet);
         void RegisterSimpleOption(IManifest mod, string optionName, string optionDesc, Func<SButton> optionGet, Action<SButton> optionSet);
+    }
+    // See https://www.nexusmods.com/stardewvalley/articles/467
+    public interface IMobilePhoneApi {
+        bool AddApp(string id, string name, Action action, Texture2D icon);
     }
 }