+++ ReleaseNotes.md
@@ -1,6 +1,13 @@
 
 # Release Notes
 
+## Version 1.1.0
+
+* If the Mobile Phone mod is installed, add a To-Dew "app" to it.
+  (This is really just an icon that brings up the same interface
+  as the hotkey.)
+
+
 ## Version 1.0.0
 
 Initial release